Trait nimbus_primitives::AccountLookup

pub trait AccountLookup<AccountId> {
    fn lookup_account(author: &NimbusId) -> Option<AccountId>;
}

A Trait to lookup runtime AccountIds from AuthorIds (probably NimbusIds) The trait is generic over the AccountId, becuase different runtimes use different notions of AccoutId. It is also generic over the AuthorId to support the usecase where the author inherent is used for beneficiary info and contains an AccountId directly.

Required Methods

fn lookup_account(author: &NimbusId) -> Option<AccountId>