Trait nimbus_primitives::NimbusApi

pub trait NimbusApi<Block: BlockT>: Core<Block> {
    fn can_author(
        &self,
        __runtime_api_at_param__: &BlockId<Block>,
        author: NimbusId,
        relay_parent: u32,
        parent_header: &Block::Header
    ) -> Result<bool, ApiError> { ... }
    fn can_author_with_context(
        &self,
        __runtime_api_at_param__: &BlockId<Block>,
        context: ExecutionContext,
        author: NimbusId,
        relay_parent: u32,
        parent_header: &Block::Header
    ) -> Result<bool, ApiError> { ... }
}

The runtime api used to predict whether a Nimbus author will be eligible in the given slot

Provided Methods

fn can_author(
&self,
__runtime_api_at_param__: &BlockId<Block>,
author: NimbusId,
relay_parent: u32,
parent_header: &Block::Header
) -> Result<bool, ApiError>

fn can_author_with_context(
&self,
__runtime_api_at_param__: &BlockId<Block>,
context: ExecutionContext,
author: NimbusId,
relay_parent: u32,
parent_header: &Block::Header
) -> Result<bool, ApiError>